+++ spec/dummy/db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160805095308) do
+ActiveRecord::Schema.define(version: 20160806012210) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -36,11 +36,13 @@ ActiveRecord::Schema.define(version: 20160805095308) do
   add_index "bloom_remit_txns", ["owner_id", "owner_type"], name: "index_bloom_remit_txns_on_owner_id_and_owner_type", using: :btree
 
   create_table "payments", force: :cascade do |t|
+    t.string "type"
   end
 
   create_table "users", force: :cascade do |t|
     t.string "email"
     t.string "bloom_remit_id"
+    t.string "type"
   end
 
 end